rewrite with typescript
