+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "preserveConstEnums": true,
+    "strictNullChecks": true,
+    "sourceMap": true,
+    "target": "es5",
+    "outDir": "./dist",
+    "moduleResolution": "node",
+    "lib": ["es2015"],
+    "rootDir": "./src"
+  }
+}
